Match-id-5466f85089067720d0ceebbffe7ce688c1361138

+++ src/core/ckpt_data_handler/feat_admit_n_evict_ckpt/feat_admit_n_evict_ckpt.cpp
@@ -136,7 +136,7 @@ void FeatAdmitNEvictCkpt::SetHistRec(string embName)
         const auto& lastTime = transArr[i + lastTimeIdxOffset];
 
         histRecs[featureId].featureId = featureId;
-        histRecs[featureId].count = count;
+        histRecs[featureId].count = static_cast<uint32_t>(count);
         histRecs[featureId].lastTime = lastTime;
         histRecs[featureId].tensorName = embName;
     }
